added serialization
